+++ src/exercicesdesignpattern/multiple/emetson/AdaptateurOie.java
@@ -3,24 +3,22 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package exercicesdesignpattern.decorator.emetson;
+package exercicesdesignpattern.multiple.emetson;
 
 /**
  *
  * @author thomas.sauvajon
  */
-public class Deco implements IEmetSon {
-    private final IEmetSon es;
+public class AdaptateurOie implements IEmetSon {
+    private Oie oie;
     
-    static int cpt;
-    
-    Deco (IEmetSon e) {
-        es = e;
+    public AdaptateurOie(Oie oie) {
+        this.oie = oie;
     }
     
     @Override
     public void emetUnSon() {
-        cpt++;
-        es.emetUnSon();
+        System.out.println("Pouet pouet");
     }
+    
 }
